Revising the cache maintenance operations
This commit comprises a series of changes to lib/vml, which are
summarized in the following.

The commit extends the cache maintenance capabilities of the earlier
function, `flush_data_cache`. By moving the barrier's implementation
into the associated header file (and making them inline), we make sure
that we do not waste cycles for function calls. This also makes sure
that we avoid reordering of the requested functionality by the compiler.

(Note that the aarch64 barriers have been adjusted to addressing the
inner shareable domain only.)

In this context, the structure of the vml library had to be adjusted
such that both aarch64 and x86 can inline the barriers in the respective
headers.

Further, this commit clearly differentiates between the `clean` and
`invalidation` functionality of the cache maintenance operations.
Besides, the changes in this commit explicitly separate data cache clean
from instruction cache invalidation functionality. All cache maintenance
functionality is limited to the `Point of Unification` (PoU).  Further
changes adjust the remaining code base to using the new functions.

bedrocksystems/vmm#319