Clean/Inv: consider the size of the written data
Problem description: hitting a breakpoint (BP), writing to the guest's
memory, and single-stepping over the hit breakpoint takes too much time.
As a result, the guest never leaves the following loop: first, it
generates an interrupt every time it tries to single-step over the hit
BP, second, it executes the IRQ handler, and, third, eventually returns
to the BP (which was intended to be skipped in the first place).

This commit avoids cleaning the data cache and invalidating the
instruction cache for the entire guest address space. The suggested
changes consider the exact position and size of the written data, when
maintaining the caches.